requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.8.0
drf-yasg==1.21.10
asgiref==3.9.1
sqlparse==0.5.3
PyYAML==6.0.2
inflection==0.5.1
packaging==25.0
pytz==2025.2
uritemplate==4.2.0
pyinstaller==6.3.0
